% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_funs.R
\name{coef.cvystat}
\alias{coef.cvystat}
\title{# cvystat deff method
#' @export
deff.cvystat <- function (object, ...) {
  if ( is.list( object ) ) object <- object[[1]]
  survey::deff( object )
}}
\usage{
\method{coef}{cvystat}(object, ...)
}
\description{
# cvystat deff method
#' @export
deff.cvystat <- function (object, ...) {
  if ( is.list( object ) ) object <- object[[1]]
  survey::deff( object )
}
}
