% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/all_funs.R
\name{complete}
\alias{complete}
\title{Complete the linearezed values with 0's}
\usage{
complete(x, ind.all)
}
\arguments{
\item{x}{vetor with linearized values for the domain}

\item{n}{the length of the vector forthe whole sample}
}
\value{
vector with 0 in the elements outside the domain
}
\description{
Complete the vector of the linearized value of an indicator for a domain with
0 values coresponding to the other domains
}
\author{
Djalma Pessoa and Anthony Damico
}
\keyword{survey}

