% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastinf.R
\name{contrastinf}
\alias{contrastinf}
\title{title here}
\usage{
contrastinf(exprlist, infunlist)
}
\arguments{
\item{exprlist}{what does exprlist need to be?}

\item{infunlist}{what does infunlist need to be?}
}
\value{
what does this function return
}
\description{
subtitle here
}
\details{
put details here
}
\examples{

# usage example here

}
\author{
Djalma Pessoa and Anthony Damico
}
\references{
do you need a references block?
}
\seealso{
do you need a seealso block?
}
\keyword{here}
\keyword{keywords}

