% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all_funs.R
\name{convey_prep}
\alias{convey_prep}
\title{prepare svydesign and svyrep.design objects for the convey package}
\usage{
convey_prep(design)
}
\arguments{
\item{design}{a survey design object of the library survey}
}
\value{
design a survey design object of the library survey
}
\description{
stores the full survey design (needed for convey functions) within the design
}
\examples{
library(vardpoor)
data(eusilc)

# linearized design: convey_prep must be run as soon as the linearized design has been created
des_eusilc <- svydesign( ids = ~rb030 , strata = ~db040 ,  weights = ~rb050 , data = eusilc )
des_eusilc <- convey_prep( des_eusilc )
# now this linearized design object is ready for analysis!

# replicate-weighted design: convey_prep must also be run as soon as the replication design has been created.
des_eusilc_rep <- survey:::as.svrepdesign( des_eusilc , type = "bootstrap" )
des_eusilc_rep <- convey_prep( des_eusilc_rep )
# now this replicate-weighted design object is ready for analysis!
}
\author{
Djalma Pessoa and Anthony Damico
}
\keyword{survey}

