% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/all_funs.R
\name{densfun}
\alias{densfun}
\title{Estimate the derivative of the cdf function using kernel estimator}
\usage{
densfun(formula, design, x, h = NULL, fun = c("F", "S"), ...)
}
\arguments{
\item{formula}{a formula specifying the income variable}

\item{design}{a design object of class \code{survey.design} of the library survey}

\item{x}{the point where the derivative is calculated}

\item{h}{value of the bandwidth based on the whole sample}

\item{if}{\code{F} estimates the derivative of the cdf function; if \code{S} estimates
the derivative of total in the tails of the distribution}
}
\value{
the value of the derivative at \code{x}
}
\description{
computes the derivative of a function in a point using kernel estimation
}
\author{
Djalma Pessoa and Anthony Damico
}
\keyword{survey}

