% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv_rules.R
\name{funs}
\alias{cl_inf}
\alias{comp_inf}
\alias{fun_par_inf}
\alias{funs}
\alias{iconst}
\alias{itot}
\alias{prod_inf}
\alias{ratio_inf}
\title{Auxiliary functions Implement the rules for Influence functions in Osier's paper}
\usage{
iconst(formula, design)

itot(formula, design)

cl_inf(a, b, big_t, big_s)

prod_inf(big_t, big_s)

ratio_inf(big_t, big_s)

comp_inf(big_t, big_s)

fun_par_inf(big_s, F, Fprime, ...)

fun_par_inf(big_s, F, Fprime, ...)
}

