% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all_funs.R
\name{icdf}
\alias{icdf}
\title{Linearization of the cdf function of a variable}
\usage{
icdf(formula, design, x, ncom, comp, ...)
}
\arguments{
\item{formula}{a formula specifying the income variable}

\item{design}{a design object of class \code{survey.design} or class \code{svyrep.design}
of the library survey}

\item{ncom}{length of the income vector for the whole sample}

\item{comp}{logical variable \code{TRUE} if the linearized variable for domains
should be completed with zeros}
}
\value{
a list with two components: the cdf estimate \code{value}
and the linearized cdf variable \code{lin}.
}
\description{
Computes the linearized variable of the cdf function in a point.
}
\examples{
library(vardpoor)
data(eusilc)
library(survey)
htot <- h_fun(eusilc$eqIncome, eusilc$rb050)
des_eusilc <- svydesign(ids=~db040, weights=~rb050, data=eusilc)
icdf_eqIncome <-icdf(~eqIncome, design=des_eusilc, 20000, ncom=nrow(eusilc), comp=TRUE)
icdf_eqIncome$value
}
\author{
Djalma Pessoa and Anthony Damico
}
\references{
Guillaume Osier (2009). Variance estimation for complex indicators
of poverty and inequality. \emph{Journal of the European Survey Research
Association}, Vol.3, No.3, pp. 167-195,
ISSN 1864-3361, URL \url{http://ojs.ub.uni-konstanz.de/srm/article/view/369}.
Jean-Claude Deville (1999). Variance estimation for complex statistics and estimators:
linearization and residual techniques. Survey Methodology, 25, 193-203,
URL \url{http://www5.statcan.gc.ca/bsolc/olc-cel/olc-cel?lang=eng&catno=12-001-X19990024882}.
}
\seealso{
\code{\link{arpr}}
}
\keyword{survey}

