% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/all_funs.R
\name{isq}
\alias{isq}
\title{Linearization of the total above a quantile or below a quantile}
\usage{
isq(formula, design, alpha, comp = TRUE, compinc, na.rm = FALSE, ...)
}
\arguments{
\item{formula}{a formula specifying the income variable}

\item{design}{a design object of class \code{survey.design} or class \code{svyrep.design}
of the library survey.  database-backed designs not supported}

\item{alpha}{the order of the quantile}

\item{compinc}{logical variable \code{TRUE} if should use the vector of sample #'values of the variable whose quantile is to be linearized}
}
\value{
Object of class "\code{cvystat}", which are vectors with a "\code{var}" attribute giving the variance and a "\code{statistic}" attribute giving the name of
the statistic.
}
\description{
Computes the linearized variable of the total in the lower tail of
the distribution of a variable.
}
\examples{
library(vardpoor)
data(eusilc)
library(survey)
des_eusilc <- svydesign(ids = ~rb030, strata =~db040,  weights = ~rb050, data = eusilc)
des_eusilc <- convey_prep( des_eusilc )
isq(~eqIncome, design=des_eusilc,compinc=TRUE, .20)
# linearized design using a variable with missings
isq( ~ py010n , design = des_eusilc, compinc=TRUE, .20 )
isq( ~ py010n , design = des_eusilc ,compinc=TRUE, .20, na.rm = TRUE )
}
\author{
Djalma Pessoa and Anthony Damico
}
\references{
Guillaume Osier (2009). Variance estimation for complex indicators
of poverty and inequality. \emph{Journal of the European Survey Research
Association}, Vol.3, No.3, pp. 167-195,
ISSN 1864-3361, URL \url{http://ojs.ub.uni-konstanz.de/srm/article/view/369}.
Jean-Claude Deville (1999). Variance estimation for complex statistics and estimators:
linearization and residual techniques. Survey Methodology, 25, 193-203,
URL \url{http://www5.statcan.gc.ca/bsolc/olc-cel/olc-cel?lang=eng&catno=12-001-X19990024882}.
}
\seealso{
\code{\link{arpr}}
}
\keyword{survey}

