% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all_funs.R
\name{SE_lin2}
\alias{SE_lin2}
\title{Extracts the se estimate}
\usage{
SE_lin2(formula, design, ...)
}
\arguments{
\item{design}{a survey design object of the library survey}

\item{object}{output of a linearizing indicator function}
}
\value{
the estime of the indicator se
}
\description{
Computes the se from the linearized variable.
}
\examples{
library(vardpoor)
data(eusilc)
library(survey)
des_eusilc <- svydesign(ids = ~rb030, strata =~db040,  weights = ~rb050, data = eusilc)
qsr_eqIncome <- svyqsr(~eqIncome, design=des_eusilc, alpha= .20)
# se estimate of isq_eqIncome for the whole sample
SE_lin2(attr(qsr_eqIncome,"lin"), des_eusilc)
}
\author{
Djalma Pessoa and Anthony Damico
}
\keyword{survey}

