% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svychu.R
\name{svychu}
\alias{svychu}
\alias{svychu.survey.design}
\alias{svychu.svyrep.design}
\alias{svychu.DBIsvydesign}
\title{CHU measure of poverty}
\usage{
svychu(formula, design, ...)

\method{svychu}{survey.design}(
  formula,
  design,
  g,
  type_thresh = "abs",
  abs_thresh = NULL,
  percent = 0.6,
  quantiles = 0.5,
  thresh = FALSE,
  na.rm = FALSE,
  deff = FALSE,
  linearized = FALSE,
  ...
)

\method{svychu}{svyrep.design}(
  formula,
  design,
  g = 0.5,
  type_thresh = "abs",
  abs_thresh = NULL,
  percent = 0.6,
  quantiles = 0.5,
  thresh = FALSE,
  na.rm = FALSE,
  deff = FALSE,
  linearized = FALSE,
  return.replicates = FALSE,
  ...
)

\method{svychu}{DBIsvydesign}(formula, design, ...)
}
\arguments{
\item{formula}{a formula specifying the income variable}

\item{design}{a design object of class \code{survey.design} or class \code{svyrep.design} from the \code{survey} library.}

\item{...}{passed to \code{svyarpr} and \code{svyarpt}}

\item{g}{If g=0 estimates the headcount ratio; If g=1 estimates the average normalised poverty gap, and if g=2 estimates the average squared normalised poverty gap}

\item{type_thresh}{type of poverty threshold. If "abs" the threshold is fixed and given the value
of abs_thresh; if "relq" it is given by percent times the quantile; if "relm" it is percent times the mean.}

\item{abs_thresh}{poverty threshold value if type_thresh is "abs"}

\item{percent}{the multiple of the the quantile or mean used in the poverty threshold definition}

\item{quantiles}{the quantile used used in the poverty threshold definition}

\item{thresh}{return the poverty threshold value}

\item{na.rm}{Should cases with missing values be dropped?}

\item{deff}{Return the design effect (see \code{survey::svymean})}

\item{linearized}{Should a matrix of linearized variables be returned}

\item{return.replicates}{Return the replicate estimates?}
}
\value{
Object of class "\code{cvystat}", which are vectors with a "\code{var}" attribute giving the variance and a "\code{statistic}" attribute giving the name of the statistic.
}
\description{
Estimate the CHU measure.
}
\details{
you must run the \code{convey_prep} function on your survey design object immediately after creating it with the \code{svydesign} or \code{svrepdesign} function.

The Clark-Hemming-Ulph class used here is mentioned in Zheng (2001) as the Chakravarty (1983) version of the CHU class, based on the ratio between the income of the poor and the poverty threshold,
using the parameter \code{g} to "weight" these ratios.
}
\examples{
library(survey)
library(laeken)
data(eusilc) ; names( eusilc ) <- tolower( names( eusilc ) )

# linearized design

des_eusilc <- svydesign( ids = ~rb030 , strata = ~db040 ,  weights = ~rb050 , data = eusilc )
des_eusilc <- convey_prep( des_eusilc )

# replicate-weighted design
des_eusilc_rep <- as.svrepdesign( des_eusilc , type = "bootstrap" )
des_eusilc_rep <- convey_prep( des_eusilc_rep )

# poverty threshold fixed
svychu(~eqincome, des_eusilc , abs_thresh=10000 , g = .5 )
# poverty threshold equal to arpt
svychu(~eqincome, des_eusilc , g = .5 , type_thresh= "relq", thresh = TRUE)
# poverty threshold equal to 0.6 times the mean
svychu(~eqincome, des_eusilc , g = .5 , type_thresh= "relm" , thresh = TRUE)
# using svrep.design:
# poverty threshold fixed
svychu(~eqincome, des_eusilc_rep , g = .5 , abs_thresh=10000)
# poverty threshold equal to arpt
svychu(~eqincome, des_eusilc_rep , g = .5 , type_thresh= "relq", thresh = TRUE)
# poverty threshold equal to 0.6 times the mean
svychu(~eqincome, des_eusilc_rep , g = .5 , type_thresh= "relm" , thresh = TRUE)

\dontrun{

# database-backed design
library(RSQLite)
library(DBI)
dbfile <- tempfile()
conn <- dbConnect( RSQLite::SQLite() , dbfile )
dbWriteTable( conn , 'eusilc' , eusilc )

dbd_eusilc <-
	svydesign(
		ids = ~rb030 ,
		strata = ~db040 ,
		weights = ~rb050 ,
		data="eusilc",
		dbname=dbfile,
		dbtype="SQLite"
	)


dbd_eusilc <- convey_prep( dbd_eusilc )

# poverty threshold fixed
svychu(~eqincome, dbd_eusilc , g = .5 , abs_thresh=10000 )
# poverty threshold equal to arpt
svychu(~eqincome, dbd_eusilc , g = .5 , type_thresh= "relq", thresh = TRUE)
# poverty threshold equal to 0.6 times the mean
svychu(~eqincome, dbd_eusilc , g = .5 , type_thresh= "relm" , thresh = TRUE)

dbRemoveTable( conn , 'eusilc' )

dbDisconnect( conn , shutdown = TRUE )

}

}
\references{
Satya R. Chakravarty (1983). A new index of poverty. \emph{Mathematical Social Sciences},
Vol. 6 , No. 3, pp. 307-313.

Stephen Clark, Richard Hemming and David Ulph (1981). On Indices
for the Measurement of Poverty. \emph{The Economic Journal}, Vol.91, No.362, (Jun., 1981), pp. 515-526,
URL \url{http://www.jstor.org/stable/2232600}.

Buhong Zheng (2001). Statistical inference for poverty measures with relative poverty lines.
\emph{Journal of Econometrics}, Vol. 101, pp. 337-356.

Vijay Verma and Gianni Betti (2011). Taylor linearization sampling errors and design effects for poverty measures
and other complex statistics. \emph{Journal Of Applied Statistics}, Vol.38, No.8, pp. 1549-1576,
URL \url{http://dx.doi.org/10.1080/02664763.2010.515674}.

Anthony B. Atkinson (1987). On the measurement of poverty.
\emph{Econometrica}, Vol.55, No.4, (Jul., 1987), pp. 749-764,
URL \url{http://www.jstor.org/stable/1911028}.

Guillaume Osier (2009). Variance estimation for complex indicators
of poverty and inequality. \emph{Journal of the European Survey Research
Association}, Vol.3, No.3, pp. 167-195,
ISSN 1864-3361, URL \url{http://ojs.ub.uni-konstanz.de/srm/article/view/369}.

Jean-Claude Deville (1999). Variance estimation for complex statistics and estimators:
linearization and residual techniques. Survey Methodology, 25, 193-203,
URL \url{http://www5.statcan.gc.ca/bsolc/olc-cel/olc-cel?lang=eng&catno=12-001-X19990024882}.
}
\seealso{
\code{\link{svyarpt}}
}
\author{
Guilherme Jacob
}
\keyword{survey}
