% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svycv.R
\name{svycv}
\alias{svycv}
\alias{svycv.DBIsvydesign}
\alias{svycv.survey.design}
\alias{svycv.svyrep.design}
\title{Coefficient of variation}
\usage{
svycv(formula, design, ...)

\method{svycv}{survey.design}(formula, design, squared = TRUE,
  na.rm = FALSE, ...)

\method{svycv}{svyrep.design}(formula, design, squared = TRUE,
  na.rm = FALSE, ...)

\method{svycv}{DBIsvydesign}(formula, design, ...)
}
\arguments{
\item{formula}{a formula specifying the income variable}

\item{design}{a design object of class \code{survey.design} or class \code{svyrep.design} from the \code{survey} library.}

\item{squared}{a logical variable defining if the result is the squared or unsquared coefficient of variation. Defaults to \code{TRUE}.}

\item{na.rm}{Should cases with missing values be dropped?}
}
\value{
Object of class "\code{cvystat}", which are vectors with a "\code{var}" attribute giving the variance and a "\code{statistic}" attribute giving the name of the statistic.
}
\description{
Estimate the coefficient of variation, a measure of inequality
}
\details{
you must run the \code{convey_prep} function on your survey design object immediately after creating it with the \code{svydesign} or \code{svrepdesign} function.
}
\examples{
library(survey)
library(vardpoor)
data(eusilc) ; names( eusilc ) <- tolower( names( eusilc ) )

# linearized design
des_eusilc <- svydesign( ids = ~rb030 , strata = ~db040 ,  weights = ~rb050 , data = eusilc )
library(convey)
des_eusilc <- convey_prep(des_eusilc)

# replicate-weighted design
des_eusilc_rep <- survey:::as.svrepdesign( des_eusilc , type = "bootstrap" )

# database-backed design
require(RSQLite)
tfile <- tempfile()
conn <- dbConnect( SQLite() , tfile )
dbWriteTable( conn , 'eusilc' , eusilc )
dbd_eusilc <- svydesign(ids = ~rb030 , strata = ~db040 ,  weights = ~rb050 , data="eusilc", dbname=tfile, dbtype="SQLite")
dbd_eusilc <- convey_prep( dbd_eusilc )

# variable no missing
svycv(~eqincome, des_eusilc, squared = TRUE, na.rm = FALSE )
svycv(~eqincome, dbd_eusilc, squared = TRUE, na.rm = FALSE )
svycv(~eqincome, des_eusilc_rep, squared = TRUE, na.rm = FALSE )

svycv(~eqincome, des_eusilc, squared = FALSE, na.rm = FALSE )
svycv(~eqincome, dbd_eusilc, squared = FALSE, na.rm = FALSE )
svycv(~eqincome, des_eusilc_rep, squared = FALSE, na.rm = FALSE )

# subset:
svycv(~eqincome, subset(des_eusilc, db040 == "Styria" ), squared = TRUE, na.rm = FALSE )
svycv(~eqincome, subset(dbd_eusilc, db040 == "Styria"), squared = TRUE, na.rm = FALSE )
svycv(~eqincome, subset(des_eusilc_rep, db040 == "Styria"), squared = TRUE, na.rm = FALSE )

svycv(~py010n, subset(des_eusilc, db040 == "Styria"), squared = FALSE, na.rm = FALSE )
svycv(~py010n, subset(dbd_eusilc, db040 == "Styria"), squared = FALSE, na.rm = FALSE )
svycv(~py010n, subset(des_eusilc_rep, db040 == "Styria"), squared = FALSE, na.rm = FALSE )

svycv(~py010n, subset(des_eusilc, db040 == "Styria"), squared = FALSE, na.rm = TRUE )
svycv(~py010n, subset(dbd_eusilc, db040 == "Styria"), squared = FALSE, na.rm = TRUE )
svycv(~py010n, subset(des_eusilc_rep, db040 == "Styria"), squared = FALSE, na.rm = TRUE )

}
\author{
Djalma Pessoa, Anthony Damico and Guilherme Jacob
}
\references{
Milorad Kovacevic and David Binder (1997). Variance Estimation for Measures of Income
Inequality and Polarization - The Estimating Equations Approach. \emph{Journal of Official Statistics},
Vol.13, No.1, 1997. pp. 41 58. URL \url{http://www.jos.nu/Articles/abstract.asp?article=13141}.
}
\seealso{
\code{\link{svygei}}
}
\keyword{survey}

