% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svyfwp.R
\name{svyfwp}
\alias{svyfwp}
\alias{svyfwp.survey.design}
\alias{svyfwp.svyrep.design}
\alias{svyfwp.DBIsvydesign}
\title{Foster-Wolfson Polarization index}
\usage{
svyfwp(formula, design, ...)

\method{svyfwp}{survey.design}(formula, design, na.rm = FALSE, deff = FALSE, influence = FALSE)

\method{svyfwp}{svyrep.design}(
  formula,
  design,
  na.rm = FALSE,
  deff = FALSE,
  influence = FALSE,
  return.replicates = FALSE
)

\method{svyfwp}{DBIsvydesign}(formula, design, ...)
}
\arguments{
\item{formula}{a formula specifying the income variable}

\item{design}{a design object of class \code{survey.design} or class \code{svyrep.design} from the \code{survey} library.}

\item{...}{future expansion}

\item{na.rm}{Should cases with missing values be dropped? Observations containing missing values in income or group variables will be dropped.}
}
\value{
Object of class "\code{cvydstat}", which are vectors with a "\code{var}" attribute giving the variance-covariance matrix and a "\code{statistic}" attribute giving the name of the statistic.
}
\description{
Estimates the Foster and Wolfson (1992) income (bi-)polarization index.
}
\details{
you must run the \code{convey_prep} function on your survey design object immediately after creating it with the \code{svydesign} or \code{svrepdesign} function.
}
\examples{
library(survey)
library(laeken)
data(eusilc) ; names( eusilc ) <- tolower( names( eusilc ) )

# linearized design
des_eusilc <- svydesign( ids = ~rb030 , strata = ~db040 ,  weights = ~rb050 , data = eusilc )
des_eusilc <- convey_prep(des_eusilc)

svyfwp( ~eqincome , design = des_eusilc )

# replicate-weighted design
des_eusilc_rep <- as.svrepdesign( des_eusilc , type = "bootstrap" )
des_eusilc_rep <- convey_prep(des_eusilc_rep)

svyfwp( ~eqincome , design = des_eusilc_rep )

\dontrun{

# linearized design using a variable with missings
svyfwp( ~ py010n , design = des_eusilc )
svyfwp( ~ py010n , design = des_eusilc , na.rm = TRUE )
# replicate-weighted design using a variable with missings
svyfwp( ~ py010n , design = des_eusilc_rep )
svyfwp( ~ py010n , design = des_eusilc_rep , na.rm = TRUE )

# database-backed design
library(RSQLite)
library(DBI)
dbfile <- tempfile()
conn <- dbConnect( RSQLite::SQLite() , dbfile )
dbWriteTable( conn , 'eusilc' , eusilc )

dbd_eusilc <-
	svydesign(
		ids = ~rb030 ,
		strata = ~db040 ,
		weights = ~rb050 ,
		data="eusilc",
		dbname=dbfile,
		dbtype="SQLite"
	)

dbd_eusilc <- convey_prep( dbd_eusilc )

svyfwp( ~ eqincome , design = dbd_eusilc )

dbRemoveTable( conn , 'eusilc' )

dbDisconnect( conn , shutdown = TRUE )

}

}
\references{
J. E. Foster and M. C. Wolfson (2010). Polarization and the decline of the middle class:
Canada and the u.S. \emph{The Journal of Economic Inequality}, Vol. 8, No. 2, p. 247-273. Originally published in 1992.

M. Kovacevic and D. A. Binder (1997). Variance Estimation for Measure of Income Inequality
and Polarization - the Estimating Equations Approach. \emph{Journal of Official Statistics}, Vol. 13, No. 1 , p. 41-58.

M. C. Wolfson (1992). When Inequalities Diverge. \emph{The American Econommic Review},
Vol. 84, No. 2, p. 353-358.
}
\author{
Guilherme Jacob
}
\keyword{survey}
