% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/svygpg.R
\name{svygpg}
\alias{svygpg}
\title{Linearization of the gender pay (wage) gap}
\usage{
svygpg(formula, design, ...)
}
\arguments{
\item{formula}{a formula specifying the gross hourly earnings variable}

\item{design}{a design object of class \code{survey.design} or class \code{svyrep.design}
of the library survey}

\item{sex}{formula with a factor with labels "male" and "female"}
}
\value{
a list with two components: the indicator estimate \code{value}
and the linearized variable \code{lin}.
}
\description{
Estimate the difference between the average gross hourly earnings of men and women expressed as a percentage of the average gross hourly earnings of men.
}
\author{
Djalma Pessoa and Anthony Damico
}
\references{
Guillaume Osier (2009). Variance estimation for complex indicators
of poverty and inequality. \emph{Journal of the European Survey Research
Association}, Vol.3, No.3, pp. 167-195,
ISSN 1864-3361, URL \url{http://ojs.ub.uni-konstanz.de/srm/article/view/369}.
Jean-Claude Deville (1999). Variance estimation for complex statistics and estimators:
linearization and residual techniques. Survey Methodology, 25, 193-203,
URL \url{http://www5.statcan.gc.ca/bsolc/olc-cel/olc-cel?lang=eng&catno=12-001-X19990024882}.
}
\seealso{
\code{\link{arpt}}
}
\keyword{survey}

