% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svyiqalpha.R
\name{svyiqalpha}
\alias{svyiqalpha}
\alias{svyiqalpha.DBIsvydesign}
\alias{svyiqalpha.survey.design}
\alias{svyiqalpha.svyrep.design}
\title{Linearization of a variable quantile}
\usage{
svyiqalpha(formula, design, ...)

\method{svyiqalpha}{survey.design}(formula, design, alpha, na.rm = FALSE, ...)

\method{svyiqalpha}{svyrep.design}(formula, design, alpha, na.rm = FALSE, ...)

\method{svyiqalpha}{DBIsvydesign}(formula, design, ...)
}
\arguments{
\item{formula}{a formula specifying the income variable}

\item{design}{a design object of class \code{survey.design} or class \code{svyrep.design} from the \code{survey} library.}

\item{alpha}{the order of the quantile}
}
\value{
Object of class "\code{cvystat}", which are vectors with a "\code{var}" attribute giving the variance and a "\code{statistic}" attribute giving the name of
the statistic.
}
\description{
Computes the linearized variable of a quantile of variable.
}
\examples{

library(vardpoor)
data(eusilc)
library(survey)
# linearized design
des_eusilc <- svydesign( ids = ~rb030 , strata = ~db040 ,  weights = ~rb050 , data = eusilc )

svyiqalpha( ~eqIncome , design = des_eusilc, .50 )

# replicate-weighted design
des_eusilc_rep <- survey:::as.svrepdesign( des_eusilc , type = "bootstrap" )

svyiqalpha( ~eqIncome , design = des_eusilc_rep, .50 )

# linearized design using a variable with missings
svyiqalpha( ~ py010n , design = des_eusilc, .50 )
svyiqalpha( ~ py010n , design = des_eusilc , .50, na.rm = TRUE )
# replicate-weighted design using a variable with missings
svyiqalpha( ~ py010n , design = des_eusilc_rep, .50 )
svyiqalpha( ~ py010n , design = des_eusilc_rep ,.50, na.rm = TRUE )

# database-backed design
require(RSQLite)
tfile <- tempfile()
conn <- dbConnect( SQLite() , tfile )
dbWriteTable( conn , 'eusilc' , eusilc )

dbd_eusilc <- svydesign(ids = ~rb030 , strata = ~db040 ,  weights = ~rb050 , data="eusilc", dbname=tfile, dbtype="SQLite")

dbd_eusilc <- convey_prep( dbd_eusilc )
svyiqalpha( ~ eqIncome , design = dbd_eusilc, .50 )

}
\author{
Djalma Pessoa and Anthony Damico
}
\references{
Guillaume Osier (2009). Variance estimation for complex indicators
of poverty and inequality. \emph{Journal of the European Survey Research
Association}, Vol.3, No.3, pp. 167-195,
ISSN 1864-3361, URL \url{http://ojs.ub.uni-konstanz.de/srm/article/view/369}.
Jean-Claude Deville (1999). Variance estimation for complex statistics and estimators:
linearization and residual techniques. Survey Methodology, 25, 193-203,
URL \url{http://www5.statcan.gc.ca/bsolc/olc-cel/olc-cel?lang=eng&catno=12-001-X19990024882}.
}
\seealso{
\code{\link{arpr}}
}
\keyword{survey}

