% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svyisq.R
\name{svyisq}
\alias{svyisq}
\alias{svyisq.DBIsvydesign}
\alias{svyisq.survey.design}
\alias{svyisq.svyrep.design}
\title{Linearization of the total below a quantile}
\usage{
svyisq(formula, design, ...)

\method{svyisq}{survey.design}(formula, design, alpha, quantile = FALSE,
  na.rm = FALSE, ...)

\method{svyisq}{svyrep.design}(formula, design, alpha, quantile = FALSE,
  na.rm = FALSE, ...)

\method{svyisq}{DBIsvydesign}(formula, design, ...)
}
\arguments{
\item{formula}{a formula specifying the income variable}

\item{design}{a design object of class \code{survey.design} or class \code{svyrep.design} from the \code{survey} library.}

\item{alpha}{the order of the quantile}

\item{quantile}{return the upper bound of the lower tail}
}
\value{
Object of class "\code{cvystat}", which are vectors with a "\code{var}" attribute giving the variance and a "\code{statistic}" attribute giving the name of
the statistic.
}
\description{
Computes the linearized variable of the total in the lower tail of
the distribution of a variable.
}
\examples{
library(vardpoor)
data(eusilc) ; names( eusilc ) <- tolower( names( eusilc ) )
library(survey)
des_eusilc <- svydesign(ids = ~rb030, strata =~db040,  weights = ~rb050, data = eusilc)
svyisq(~eqincome, design=des_eusilc,.20 , quantile = TRUE)
# replicate-weighted design
des_eusilc_rep <- survey:::as.svrepdesign( des_eusilc , type = "bootstrap" )
svyisq( ~eqincome , design = des_eusilc_rep, .20 , quantile = TRUE )
# linearized design using a variable with missings
svyisq( ~ py010n , design = des_eusilc, .20 )
svyisq( ~ py010n , design = des_eusilc , .20, na.rm = TRUE )
# replicate-weighted design using a variable with missings
svyisq( ~ py010n , design = des_eusilc_rep, .20 )
svyisq( ~ py010n , design = des_eusilc_rep , .20,  na.rm = TRUE )

# database-backed design
require(RSQLite)
tfile <- tempfile()
conn <- dbConnect( SQLite() , tfile )
dbWriteTable( conn , 'eusilc' , eusilc )

dbd_eusilc <- svydesign(ids = ~rb030 , strata = ~db040 ,  weights = ~rb050 , data="eusilc", dbname=tfile, dbtype="SQLite")

dbd_eusilc <- convey_prep( dbd_eusilc )
svyisq( ~ eqincome , design = dbd_eusilc, .20 )

}
\author{
Djalma Pessoa and Anthony Damico
}
\references{
Guillaume Osier (2009). Variance estimation for complex indicators
of poverty and inequality. \emph{Journal of the European Survey Research
Association}, Vol.3, No.3, pp. 167-195,
ISSN 1864-3361, URL \url{http://ojs.ub.uni-konstanz.de/srm/article/view/369}.
Jean-Claude Deville (1999). Variance estimation for complex statistics and estimators:
linearization and residual techniques. Survey Methodology, 25, 193-203,
URL \url{http://www5.statcan.gc.ca/bsolc/olc-cel/olc-cel?lang=eng&catno=12-001-X19990024882}.
}
\seealso{
\code{\link{arpr}}
}
\keyword{survey}

