% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svykmoi.R
\name{svykmoi}
\alias{svykmoi}
\alias{svykmoi.survey.design}
\alias{svykmoi.svyrep.design}
\alias{svykmoi.DBIsvydesign}
\title{Kobus and Milos (2012) ordinal inequality measure (EXPERIMENTAL)}
\usage{
svykmoi(formula, design, ...)

\method{svykmoi}{survey.design}(formula, design, ref.point = NULL, a = 1,
  b = 1, na.rm = FALSE, ...)

\method{svykmoi}{svyrep.design}(formula, design, ref.point = NULL, a = 1,
  b = 1, na.rm = FALSE, ...)

\method{svykmoi}{DBIsvydesign}(formula, design, ...)
}
\arguments{
\item{formula}{a formula specifying the income variable}

\item{design}{a design object of class \code{survey.design} or class \code{svyrep.design} from the \code{survey} library.}

\item{...}{future expansion}

\item{ref.point}{a string giving the labels of the "reference point". If not given, it defaults to the median category.}

\item{a}{a parameter giving the sensivity towards inequality below the reference point. Default to 1.}

\item{b}{a parameter giving the sensivity towards inequality over the reference point. Default to 1.}

\item{na.rm}{Should cases with missing values be dropped?}
}
\value{
Object of class "\code{cvystat}", which are vectors with a "\code{var}" attribute giving the variance and a "\code{statistic}" attribute giving the name of the statistic.
}
\description{
Estimate the Kobus and Milos (2012) inequality measure, a measure of inequality for ordinal variables.
}
\details{
you must run the \code{convey_prep} function on your survey design object immediately after creating it with the \code{svydesign} or \code{svrepdesign} function.

This measure only allows for strictly positive variables.
}
\note{
This function is experimental and is subject to change in later versions.
}
\examples{
library(survey)
library(vardpoor)
data(ses) ; names( ses ) <- tolower( names( ses ) )

# linearized design
des_ses <- svydesign( ids = ~rb030 , strata = ~db040 ,  weights = ~rb050 , data = ses )
des_ses <- convey_prep(des_ses)
des_ses <- update( des_ses, education = ordered( education ) )

svykmoi( ~education , design = des_ses , ref.point = "ISCED 3 and 4" )

# replicate-weighted design
des_ses_rep <- as.svrepdesign( des_ses , type = "bootstrap" )
des_ses_rep <- convey_prep(des_ses_rep)

svykmoi( ~education , design = des_ses_rep , ref.point = "ISCED 3 and 4" )

\dontrun{

# database-backed design
library(MonetDBLite)
library(DBI)
dbfolder <- tempdir()
conn <- dbConnect( MonetDBLite::MonetDBLite() , dbfolder )
dbWriteTable( conn , 'ses' , ses )

dbd_ses <-
	svydesign(
		ids = ~rb030 ,
		strata = ~db040 ,
		weights = ~rb050 ,
		data="ses",
		dbname=dbfolder,
		dbtype="MonetDBLite"
	)

dbd_ses <- convey_prep( dbd_ses )
dbd_ses <- update( dbd_ses, education = ordered( education ) )

svykmoi( ~education , design = dbd_ses , ref.point = "ISCED 3 and 4" )

dbRemoveTable( conn , 'ses' )

dbDisconnect( conn , shutdown = TRUE )

}

}
\references{
Martyna Kobus and Pyotr Milos. Inequality decomposition by population subgroups for ordinal data.
\emph{Journal of Health Economics}, Vol.31, No.1 (2016), pp. 15-21.
URL \url{http://www.sciencedirect.com/science/article/pii/S0167629611001664}.
}
\seealso{
\code{\link{svygei}}
}
\author{
Guilherme Jacob
}
\keyword{survey}
