% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/svyrmir.R
\name{svyrmir}
\alias{svyrmir}
\alias{svyrmir.DBIsvydesign}
\alias{svyrmir.survey.design}
\alias{svyrmir.svyrep.design}
\title{Relative median income ratio}
\usage{
svyrmir(formula, design, ...)

\method{svyrmir}{survey.design}(formula, design, age, agelim, order)

\method{svyrmir}{svyrep.design}(formula, design, order = 0.5, age, agelim,
  ...)

\method{svyrmir}{DBIsvydesign}(x, design, ...)
}
\arguments{
\item{formula}{a formula specifying the income variable}

\item{design}{a design object of class \code{survey.design} or class \code{svyrep.design}
of the library survey}

\item{age}{formula defining the variable age}

\item{agelim}{the age cutpoint, the default is 65}

\item{order}{income quantile order, usually .5}
}
\value{
Object of class "\code{cvystat}", which are vectors with a "\code{var}" attribute giving the variance and a "\code{statistic}" attribute giving the name of the statistic.
}
\description{
Estimates the ratio between the median income of people with age above 65 and the
median income of people with age below 65.
}
\examples{
library(vardpoor)
data(eusilc)
library(survey)
des_eusilc <- svydesign(ids=~db040, weights=~rb050, data=eusilc)
des_eusilc <- convey_prep( des_eusilc )
rmir_eusilc<- svyrmir(~eqIncome, des_eusilc, age=~age, agelim = 65, order=.50 )
}
\author{
Djalma Pessoa and Anthony Damico
#'
}
\references{
Guillaume Osier (2009). Variance estimation for complex indicators
of poverty and inequality. \emph{Journal of the European Survey Research
Association}, Vol.3, No.3, pp. 167-195,
ISSN 1864-3361, URL \url{http://ojs.ub.uni-konstanz.de/srm/article/view/369}.
Jean-Claude Deville (1999). Variance estimation for complex statistics and estimators:
linearization and residual techniques. Survey Methodology, 25, 193-203,
URL \url{http://www5.statcan.gc.ca/bsolc/olc-cel/olc-cel?lang=eng&catno=12-001-X19990024882}.
}
\seealso{
\code{\link{arpt}}
}
\keyword{survey}

