% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svyScurve.R
\name{svyScurve}
\alias{svyScurve}
\alias{svyScurve.DBIsvydesign}
\alias{svyScurve.survey.design}
\alias{svyScurve.svyrep.design}
\title{Allison-Foster (2004) S-curve}
\usage{
svyScurve(formula, design, ...)

\method{svyScurve}{survey.design}(formula, design, add = FALSE,
  line.color = "black", alpha = 0.05, na.rm = TRUE, ...)

\method{svyScurve}{svyrep.design}(formula, design, add = FALSE,
  line.color = "black", alpha = 0.05, na.rm = TRUE, ...)

\method{svyScurve}{DBIsvydesign}(formula, design, ...)
}
\arguments{
\item{formula}{a formula specifying the variable. Variable has to be ordered factor.}

\item{design}{a design object of class \code{survey.design} or class \code{svyrep.design} from the \code{survey} library.}

\item{...}{future expansion}

\item{na.rm}{Should cases with missing values be dropped?}

\item{curve}{should the S-curve be plotted? Defaults to \code{TRUE}.}
}
\value{
Object of class "\code{cvystat}", which are vectors with a "\code{var}" attribute giving the variance and a "\code{statistic}" attribute giving the name of the statistic.
}
\description{
Plot the S-curve, a graph for inequality analysis of ordinal variables.
}
\details{
you must run the \code{convey_prep} function on your survey design object immediately after creating it with the \code{svydesign} or \code{svrepdesign} function.
}
\examples{

\dontrun{

# database-backed design
library(MonetDBLite)
library(DBI)
dbfolder <- tempdir()
conn <- dbConnect( MonetDBLite::MonetDBLite() , dbfolder )
dbWriteTable( conn , 'eusilc' , eusilc )

dbd_eusilc <-
	svydesign(
		ids = ~rb030 ,
		strata = ~db040 ,
		weights = ~rb050 ,
		data="eusilc",
		dbname=dbfolder,
		dbtype="MonetDBLite"
	)

dbd_eusilc <- convey_prep( dbd_eusilc )

# linearized
svyScurve(~eqincome+hy050n, design = dbd_eusilc, alpha = .5, beta = .5, na.rm = FALSE )
svyScurve(~eqincome+hy050n, design = dbd_eusilc, alpha = .5, beta = 0, na.rm = FALSE )
svyScurve(~eqincome+hy050n, design = dbd_eusilc, alpha = .5, beta = -.5, na.rm = FALSE )
svyScurve(~eqincome+hy050n, design = dbd_eusilc, alpha = .5, beta = -1, na.rm = FALSE )
svyScurve(~eqincome+hy050n, design = dbd_eusilc, alpha = .5, beta = -2, na.rm = FALSE )

# subsetting:
sub_dbd_eusilc <- subset( dbd_eusilc, db040 == "Styria")
svyScurve(~eqincome+hy050n, design = sub_dbd_eusilc, alpha = .5, beta = .5, na.rm = FALSE )

dbRemoveTable( conn , 'eusilc' )

dbDisconnect( conn , shutdown = TRUE )

}

}
\author{
Guilherme Jacob
}
\references{
R. Andrew Allison and James E. Foster (2007). Measuring health inequality using qualitative data.
Journal of Health Economics, v. 23, n. 3, May 2004, pp. 505-524.
}
\seealso{
\code{\link{svyfgt}}
}
\keyword{survey}

