% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svysen.R
\name{svysen}
\alias{svysen}
\alias{svysen.DBIsvydesign}
\alias{svysen.survey.design}
\alias{svysen.svyrep.design}
\title{Sen (1976) poverty index}
\usage{
svysen(formula, design, ...)

\method{svysen}{survey.design}(formula, design, type_thresh = "abs",
  abs_thresh = NULL, percent = 0.6, quantiles = 0.5, na.rm = FALSE,
  thresh = FALSE, ...)

\method{svysen}{svyrep.design}(formula, design, type_thresh = "abs",
  abs_thresh = NULL, percent = 0.6, quantiles = 0.5, na.rm = FALSE,
  thresh = FALSE, ...)

\method{svysen}{DBIsvydesign}(formula, design, ...)
}
\arguments{
\item{formula}{a formula specifying the income variable}

\item{design}{a design object of class \code{survey.design} or class \code{svyrep.design} from the \code{survey} library.}

\item{...}{passed to \code{svyarpr} and \code{svyarpt}}

\item{type_thresh}{type of poverty threshold. If "abs" the threshold is fixed and given the value
of abs_thresh; if "relq" it is given by percent times the quantile; if "relm" it is percent times the mean.}

\item{abs_thresh}{poverty threshold value if type_thresh is "abs"}

\item{percent}{the multiple of the the quantile or mean used in the poverty threshold definition}

\item{quantiles}{the quantile used used in the poverty threshold definition}

\item{na.rm}{Should cases with missing values be dropped?}

\item{thresh}{return the poverty threshold value}
}
\value{
Object of class "\code{cvystat}", which are vectors with a "\code{var}" attribute giving the variance and a "\code{statistic}" attribute giving the name of the statistic.
}
\description{
Estimate the Sen (1976) poverty measure.
}
\details{
you must run the \code{convey_prep} function on your survey design object immediately after creating it with the \code{svydesign} or \code{svrepdesign} function.
}
\examples{
library(survey)
library(vardpoor)
data(eusilc) ; names( eusilc ) <- tolower( names( eusilc ) )

# linearized design

des_eusilc <- svydesign( ids = ~rb030 , strata = ~db040 ,  weights = ~rb050 , data = eusilc )
des_eusilc <- convey_prep( des_eusilc )

# replicate-weighted design
des_eusilc_rep <- as.svrepdesign( des_eusilc , type = "bootstrap" )
des_eusilc_rep <- convey_prep( des_eusilc_rep )

# poverty threshold fixed
svysen(~eqincome, des_eusilc, abs_thresh=10000)
# poverty threshold equal to arpt
svysen(~eqincome, des_eusilc, type_thresh= "relq" , thresh = TRUE)
# poverty threshold equal to .6 times the mean
svysen(~eqincome, des_eusilc, type_thresh= "relm", thresh = TRUE)

#  using svrep.design:
# poverty threshold fixed
svysen(~eqincome, des_eusilc_rep, abs_thresh=10000)
# poverty threshold equal to arpt
svysen(~eqincome, des_eusilc_rep, type_thresh= "relq" , thresh = TRUE)
# poverty threshold equal to .6 times the mean
svysen(~eqincome, des_eusilc_rep, type_thresh= "relm" , thresh = TRUE)


\dontrun{

# database-backed design
library(MonetDBLite)
library(DBI)
dbfolder <- tempdir()
conn <- dbConnect( MonetDBLite::MonetDBLite() , dbfolder )
dbWriteTable( conn , 'eusilc' , eusilc )

dbd_eusilc <-
	svydesign(
		ids = ~rb030 ,
		strata = ~db040 ,
		weights = ~rb050 ,
		data="eusilc",
		dbname=dbfolder,
		dbtype="MonetDBLite"
	)


dbd_eusilc <- convey_prep( dbd_eusilc )

# headcount ratio, poverty threshold fixed
svysen(~eqincome, dbd_eusilc, abs_thresh=10000)
# poverty gap index, poverty threshold fixed
svysen(~eqincome, dbd_eusilc, abs_thresh=10000)
# headcount ratio, poverty threshold equal to arpt
svysen(~eqincome, dbd_eusilc, type_thresh= "relq", thresh = TRUE)
# poverty gap index, poverty threshold equal to arpt
svysen(~eqincome, dbd_eusilc, type_thresh= "relq")
# headcount ratio, poverty threshold equal to .6 times the mean
svysen(~eqincome, dbd_eusilc, type_thresh= "relm")
# poverty gap index, poverty threshold equal to 0.6 times the mean
svysen(~eqincome, dbd_eusilc, type_thresh= "relm")

dbRemoveTable( conn , 'eusilc' )

dbDisconnect( conn , shutdown = TRUE )

}

}
\author{
Guilherme Jacob, Djalma Pessoa and Anthony Damico
}
\references{
Amartya Sen (2009). Poverty: An Ordinal Approach to Measurement.
\emph{Econometric}, v. 44, n. 3, pp. 219-231.
URL \url{http://www.jstor.org/stable/1912718}.
}
\seealso{
\code{\link{svyarpt}}
}
\keyword{survey}

