% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all_funs.R
\name{convey_prep}
\alias{convey_prep}
\title{prepare linearized svydesign objects for convey package}
\usage{
convey_prep(design)
}
\arguments{
\item{design}{a survey design object of the library survey}
}
\value{
design a survey design object of the library survey
}
\description{
stores the full survey design (needed for linearized convey functions) within the design
}
\examples{
library(vardpoor)
data(eusilc)
des_eusilc <- svydesign(ids=~db040, weights=~rb050, data=eusilc)
des_eusilc <- convey_prep( des_eusilc )
}
\author{
Djalma Pessoa and Anthony Damico
}
\keyword{survey}

