% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all_funs.R
\name{complete}
\alias{SE_lin}
\alias{cl_inf}
\alias{complete}
\alias{computeQuantiles}
\alias{densfun}
\alias{h_fun}
\alias{icdf}
\alias{iqalpha}
\alias{isq}
\alias{itot}
\alias{prod_inf}
\alias{ratio_inf}
\title{Complete the linearezed values with 0's}
\usage{
complete(x, n)

h_fun(inc_var, w)

densfun(formula, design, x, htot = NULL, fun = c("F", "S"), ...)

icdf(formula, design, x, ncom, comp, ...)

iqalpha(formula, design, alpha, h = NULL, ncom, comp, incvec = NULL, ...)

isq(formula, design, alpha, type = c("inf", "sup"), h = NULL, ncom, incvec,
  ...)

computeQuantiles(xx, w, p = quantiles)

SE_lin(object, design)

itot(formula, design)

cl_inf(a, b, T, S)

prod_inf(T, S)

ratio_inf(T, S)
}
\arguments{
\item{x}{vetor with linearized values for the domain}

\item{n}{the length of the vector forthe whole sample}

\item{inc_var}{income variable used in the estimation of the indicators}

\item{w}{vector of design weights}

\item{formula}{a formula specifying the income variable}

\item{design}{a design object of class \code{survey.design} of the library survey}

\item{htot}{value of the bandwidth based on the whole sample}

\item{ncom}{length of the income vector for the whole sample}

\item{comp}{logical variable \code{TRUE} if the linearized variable for domains
should be completed with zeros}

\item{alpha}{the order of the quantile}

\item{h}{bandwidth to estimate the derivative of the total in the tails}

\item{type}{\code{inf} for the lower tail and \code{sup} for the upper tail}

\item{object}{output of a linearizing indicator function}

\item{x}{the point where the derivative is calculated}

\item{if}{\code{F} estimates the derivative of the cdf function; if \code{S} estimates
the derivative of total in the tails of the distribution}

\item{formula}{a formula specifying the income variable}

\item{design}{a design object of class \code{survey.design} or class \code{svyrep.design}
of the library survey}

\item{formula}{a formula specifying the income variable}

\item{design}{a design object of class \code{survey.design} or class \code{svyrep.design}
of the library survey}

\item{ncom}{length of the income vector for the whole sample}

\item{comp}{logical variable \code{TRUE} if the linearized variable for domains
should be completed with zeros}

\item{invec}{vector of sample values of the variable whose quantile is to be linearized}

\item{formula}{a formula specifying the income variable}

\item{design}{a design object of class \code{survey.design} or class \code{svyrep.design}
of the library survey}

\item{alpha}{the order of the quantile}

\item{ncom}{length of the income vector for the whole sample}

\item{invec}{vector of sample values of the variable whose quantile is to be linearized}

\item{design}{a survey design object of the library survey}
}
\value{
vector with 0 in the elements outside the domain

value of the bandwidth

the value of the derivative at \code{x}

a list with two components: the cdf estimate \code{value}
and the linearized cdf variable \code{lin}.

a list with two components: the quantile estimate \code{value}
and the linearized quantile variable \code{lin}.

a list with two components: the tail total estimate \code{value}
and the linearized of the tail total \code{lin}.

the estime of the indicator se
}
\description{
Complete the vector of the linearized value of an indicator for a domain with
0 values coresponding to the other domains

Using the whole sample, computes the bandwith used to get the linearized variable

computes the derivative of a function in a point using kernel estimation

Computes the linearized variable of the cdf function in a point.

Computes the linearized variable of a quantile of variable.

Computes the linearized variable of the total in the lower or upper tail of
the distribution of a variable.

Computes the se from the linearized variable.
}
\examples{
library(vardpoor)
data(eusilc)
library(survey)
htot <- h_fun(eusilc$eqIncome, eusilc$rb050)
des_eusilc <- svydesign(ids=~db040, weights=~rb050, data=eusilc)
icdf_eqIncome <-icdf(~eqIncome, design=des_eusilc, 20000, ncom=nrow(eusilc), comp=TRUE)
icdf_eqIncome$value
library(vardpoor)
data(eusilc)
library(survey)
htot <- h_fun(eusilc$eqIncome, eusilc$rb050)
des_eusilc <- svydesign(ids=~db040, weights=~rb050, data=eusilc)
iqalpha_eqIncome <-iqalpha(~eqIncome, design=des_eusilc, .50, ncom=nrow(eusilc),
comp=TRUE, eusilc$eqIncome )
iqalpha_eqIncome$value
library(vardpoor)
data(eusilc)
library(survey)
htot <- h_fun(eusilc$eqIncome, eusilc$rb050)
des_eusilc <- svydesign(ids=~db040, weights=~rb050, data=eusilc)
isq_eqIncome <-isq(~eqIncome, design=des_eusilc, .80, 'sup', htot,
ncom=nrow(eusilc),  eusilc$eqIncome)
isq_eqIncome$value
library(vardpoor)
data(eusilc)
library(survey)
htot <- h_fun(eusilc$eqIncome, eusilc$rb050)
des_eusilc <- svydesign(ids=~db040, weights=~rb050, data=eusilc)
qsr_eqIncome <- svyqsr(~eqIncome, design=des_eusilc, alpha= .20, ncom = nrow(eusilc),
comp=TRUE, incvec = eusilc$eqIncome)
# se estimate of isq_eqIncome for the whole sample
SE_lin(qsr_eqIncome, des_eusilc)
# se estimates for domains
isq_eqIncome_dom <-  svyby(~eqIncome, by= ~db040, design=des_eusilc,
FUN=svyqsr, alpha=.20,  h= htot, ncom=nrow(des_eusilc$variables), comp=TRUE,
incvec= eusilc$eqIncome, deff=FALSE, keep.var=FALSE)
SE_lin(isq_eqIncome_dom, des_eusilc)
}
\author{
Djalma Pessoa and Anthony Damico

Djalma Pessoa and Anthony Damico

Djalma Pessoa and Anthony Damico

Djalma Pessoa and Anthony Damico

Djalma Pessoa and Anthony Damico

Djalma Pessoa and Anthony Damico

Djalma Pessoa and Anthony Damico
}
\references{
Guillaume Osier (2009). Variance estimation for complex indicators
of poverty and inequality. \emph{Journal of the European Survey Research
Association}, Vol.3, No.3, pp. 167-195,
ISSN 1864-3361, URL \url{http://ojs.ub.uni-konstanz.de/srm/article/view/369}.
Jean-Claude Deville (1999). Variance estimation for complex statistics and estimators:
linearization and residual techniques. Survey Methodology, 25, 193-203,
URL \url{http://www5.statcan.gc.ca/bsolc/olc-cel/olc-cel?lang=eng&catno=12-001-X19990024882}.

Guillaume Osier (2009). Variance estimation for complex indicators
of poverty and inequality. \emph{Journal of the European Survey Research
Association}, Vol.3, No.3, pp. 167-195,
ISSN 1864-3361, URL \url{http://ojs.ub.uni-konstanz.de/srm/article/view/369}.
Jean-Claude Deville (1999). Variance estimation for complex statistics and estimators:
linearization and residual techniques. Survey Methodology, 25, 193-203,
URL \url{http://www5.statcan.gc.ca/bsolc/olc-cel/olc-cel?lang=eng&catno=12-001-X19990024882}.

Guillaume Osier (2009). Variance estimation for complex indicators
of poverty and inequality. \emph{Journal of the European Survey Research
Association}, Vol.3, No.3, pp. 167-195,
ISSN 1864-3361, URL \url{http://ojs.ub.uni-konstanz.de/srm/article/view/369}.
Jean-Claude Deville (1999). Variance estimation for complex statistics and estimators:
linearization and residual techniques. Survey Methodology, 25, 193-203,
URL \url{http://www5.statcan.gc.ca/bsolc/olc-cel/olc-cel?lang=eng&catno=12-001-X19990024882}.
}
\seealso{
\code{\link{arpr}}

\code{\link{arpr}}

\code{\link{arpr}}
}
\keyword{survey}

