% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/deriv_rules.R
\name{funs}
\alias{cl_inf}
\alias{comp_inf}
\alias{contrastinf}
\alias{fun_par_inf}
\alias{funs}
\alias{iconst}
\alias{itot}
\alias{prod_inf}
\alias{ratio_inf}
\title{Auxiliary functions Implement the rules for Influence functions in Osier's paper}
\usage{
iconst(formula, design)

itot(formula, design)

cl_inf(a, b, T, S)

prod_inf(T, S)

ratio_inf(T, S)

comp_inf(T, S)

fun_par_inf(S, F, Fprime, ...)

contrastinf(exprlist, infunlist)

fun_par_inf(S, F, Fprime, ...)
}
\description{
Auxiliary functions Implement the rules for Influence functions in Osier's paper
}

