% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_funs.R
\name{h_fun}
\alias{h_fun}
\title{Computes the bandwidth needed to compute the derivative of the cdf function}
\usage{
h_fun(incvar, w)
}
\arguments{
\item{incvar}{income variable used in the estimation of the indicators}

\item{w}{vector of design weights}
}
\value{
value of the bandwidth
}
\description{
Using the whole sample, computes the bandwith used to get the linearized variable
}
\author{
Djalma Pessoa and Anthony Damico
}
\keyword{survey}
