% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/all_funs.R
\name{isq}
\alias{isq}
\title{Linearization of the total above a quantile or below a quantile}
\usage{
isq(formula, design, alpha, type = c("inf", "sup"), h = NULL, ncom, incvec,
  ...)
}
\arguments{
\item{formula}{a formula specifying the income variable}

\item{design}{a design object of class \code{survey.design} or class \code{svyrep.design}
of the library survey}

\item{alpha}{the order of the quantile}

\item{type}{\code{inf} for the lower tail and \code{sup} for the upper tail}

\item{h}{bandwidth to estimate the derivative of the total in the tails}

\item{ncom}{length of the income vector for the whole sample}

\item{invec}{vector of sample values of the variable whose quantile is to be linearized}
}
\value{
a list with two components: the tail total estimate \code{value}
and the linearized of the tail total \code{lin}.
}
\description{
Computes the linearized variable of the total in the lower or upper tail of
the distribution of a variable.
}
\examples{
library(vardpoor)
data(eusilc)
library(survey)
htot <- h_fun(eusilc$eqIncome, eusilc$rb050)
des_eusilc <- svydesign(ids=~db040, weights=~rb050, data=eusilc)
isq_eqIncome <-isq(~eqIncome, design=des_eusilc, .80, "sup", htot,
ncom=nrow(eusilc),  eusilc$eqIncome)
isq_eqIncome$value
}
\author{
Djalma Pessoa and Anthony Damico
}
\references{
Guillaume Osier (2009). Variance estimation for complex indicators
of poverty and inequality. \emph{Journal of the European Survey Research
Association}, Vol.3, No.3, pp. 167-195,
ISSN 1864-3361, URL \url{http://ojs.ub.uni-konstanz.de/srm/article/view/369}.
Jean-Claude Deville (1999). Variance estimation for complex statistics and estimators:
linearization and residual techniques. Survey Methodology, 25, 193-203,
URL \url{http://www5.statcan.gc.ca/bsolc/olc-cel/olc-cel?lang=eng&catno=12-001-X19990024882}.
}
\seealso{
\code{\link{arpr}}
}
\keyword{survey}

