% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/all_funs.R
\name{SE_lin2}
\alias{SE_lin2}
\title{Extracts the se estimate}
\usage{
SE_lin2(object, design)
}
\arguments{
\item{object}{output of a linearizing indicator function}

\item{design}{a survey design object of the library survey}
}
\value{
the estime of the indicator se
}
\description{
Computes the se from the linearized variable.
}
\examples{
library(vardpoor)
data(eusilc)
library(survey)
htot <- h_fun(eusilc$eqIncome, eusilc$rb050)
des_eusilc <- svydesign(ids=~db040, weights=~rb050, data=eusilc)
qsr_eqIncome <- svyqsr(~eqIncome, design=des_eusilc, alpha= .20, ncom = rownames(eusilc),
comp=TRUE, incvec = eusilc$eqIncome)
# se estimate of isq_eqIncome for the whole sample
SE_lin2(qsr_eqIncome, des_eusilc)
# se estimates for domains
isq_eqIncome_dom <-  svyby(~eqIncome, by= ~db040, design=des_eusilc,
FUN=svyqsr, alpha=.20,  h= htot, ncom=rownames(eusilc), comp=TRUE,
incvec= eusilc$eqIncome, deff=FALSE, keep.var=FALSE)
SE_lin2(isq_eqIncome_dom, des_eusilc)
}
\author{
Djalma Pessoa and Anthony Damico
}
\keyword{survey}

