% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/svyarpr.R
\name{svyarpr}
\alias{svyarpr}
\alias{svyarpr.DBIsvydesign}
\alias{svyarpr.survey.design}
\alias{svyarpr.svyrep.design}
\title{At-risk-of-poverty rate}
\usage{
svyarpr(formula, design, ...)

\method{svyarpr}{survey.design}(formula, design, order = 0.5, percent = 0.6,
  comp = TRUE, ...)

\method{svyarpr}{svyrep.design}(formula, design, order = 0.5, percent = 0.6,
  ...)

\method{svyarpr}{DBIsvydesign}(x, design, ...)
}
\arguments{
\item{formula}{a formula specifying the income variable}

\item{design}{a design object of class \code{survey.design} or class \code{svyrep.design}
of the library survey}

\item{order}{income quantile order, usually .5}

\item{percent}{fraction of the quantile, usually .60}

\item{comp}{logical variable \code{TRUE} if the linearized variable for domains
should be completed with zeros}
}
\value{
Object of class "\code{cvystat}", which are vectors with a "\code{var}" attribute giving the variance and a "\code{statistic}" attribute giving the name of the statistic.
}
\description{
Estimate the proportion of persons with income below the
at-risk-of-poverty threshold.
}
\examples{
library(vardpoor)
data(eusilc)
library(survey)
des_eusilc <- svydesign(ids=~db040, weights=~rb050, data=eusilc)
des_eusilc <- convey_prep( des_eusilc )
arpr_eqIncome<- svyarpr(~eqIncome, design=des_eusilc, .5, .6, comp=TRUE)
des_eusilc_rep<- as.svrepdesign(des_eusilc, type="bootstrap")
svyarpr(~eqIncome, des_eusilc_rep, 0.5, 0.6)
}
\author{
Djalma Pessoa and Anthony Damico
#'
}
\references{
Guillaume Osier (2009). Variance estimation for complex indicators
of poverty and inequality. \emph{Journal of the European Survey Research
Association}, Vol.3, No.3, pp. 167-195,
ISSN 1864-3361, URL \url{http://ojs.ub.uni-konstanz.de/srm/article/view/369}.
Jean-Claude Deville (1999). Variance estimation for complex statistics and estimators:
linearization and residual techniques. Survey Methodology, 25, 193-203,
URL \url{http://www5.statcan.gc.ca/bsolc/olc-cel/olc-cel?lang=eng&catno=12-001-X19990024882}.
}
\seealso{
\code{\link{arpt}}
}
\keyword{survey}

