% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svyarpt.R
\name{svyarpt}
\alias{svyarpt}
\alias{svyarpt.DBIsvydesign}
\alias{svyarpt.survey.design}
\alias{svyarpt.svyrep.design}
\title{At-risk-of-poverty threshold}
\usage{
svyarpt(formula, design, ...)

\method{svyarpt}{survey.design}(formula, design, order = 0.5, percent = 0.6,
  comp = TRUE, na.rm = FALSE, ...)

\method{svyarpt}{svyrep.design}(formula, design, order = 0.5, percent = 0.6,
  na.rm = FALSE, ...)

\method{svyarpt}{DBIsvydesign}(x, design, ...)
}
\arguments{
\item{formula}{a formula specifying the income variable}

\item{design}{a design object of class \code{survey.design} or class \code{svyrep.design} of the library survey.  database-backed designs not supported}

\item{order}{income quantile order, usually .50 (median)}

\item{percent}{fraction of the quantile, usually .60}

\item{comp}{logical variable \code{TRUE} if the inearized variable for domains should be completed with zeros}

\item{na.rm}{Should cases with missing values be dropped?}
}
\value{
Object of class "\code{cvystat}", which are vectors with a "\code{var}" attribute giving the variance and a "\code{statistic}" attribute giving the name of the statistic.
}
\description{
The standard definition is to use 60\% of the median income.
}
\details{
you must run the \code{convey_prep} function on your survey design object immediately after creating it with the \code{svydesign} or \code{svrepdesign} function.
}
\examples{

library(survey)
library(vardpoor)
data(eusilc)

# linearized design

des_eusilc <- svydesign( ids = ~rb030 , strata = ~db040 ,  weights = ~rb050 , data = eusilc )
des_eusilc <- convey_prep( des_eusilc )

svyarpt( ~eqIncome , design = des_eusilc )

# replicate-weighted design
des_eusilc_rep <- survey:::as.svrepdesign( des_eusilc , type = "bootstrap" )
des_eusilc_rep <- convey_prep( des_eusilc_rep )
svyarpt( ~eqIncome , design = des_eusilc_rep )

# linearized design using a variable with missings
svyarpt( ~ py010n , design = des_eusilc )
svyarpt( ~ py010n , design = des_eusilc , na.rm = TRUE )
# replicate-weighted design using a variable with missings
svyarpt( ~ py010n , design = des_eusilc_rep )
svyarpt( ~ py010n , design = des_eusilc_rep , na.rm = TRUE )


# database-backed design
require(RSQLite)
tfile <- tempfile()
conn <- dbConnect( SQLite() , tfile )
dbWriteTable( conn , 'eusilc' , eusilc )

dbd_eusilc <- svydesign(ids = ~rb030 , strata = ~db040 ,  weights = ~rb050 , data="eusilc", dbname=tfile, dbtype="SQLite")
dbd_eusilc <- convey_prep( dbd_eusilc )

svyarpt( ~ eqIncome , design = dbd_eusilc )

}
\author{
Djalma Pessoa and Anthony Damico
}
\references{
Guillaume Osier (2009). Variance estimation for complex indicators
of poverty and inequality. \emph{Journal of the European Survey Research
Association}, Vol.3, No.3, pp. 167-195,
ISSN 1864-3361, URL \url{http://ojs.ub.uni-konstanz.de/srm/article/view/369}.
Jean-Claude Deville (1999). Variance estimation for complex statistics and estimators:
linearization and residual techniques. Survey Methodology, 25, 193-203,
URL \url{http://www5.statcan.gc.ca/bsolc/olc-cel/olc-cel?lang=eng&catno=12-001-X19990024882}.
}
\seealso{
\code{\link{arpr}}
}
\keyword{survey}

