% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svygini.R
\name{svyatk}
\alias{svyatk}
\alias{svyatk.DBIsvydesign}
\alias{svyatk.survey.design}
\alias{svyatk.svyrep.design}
\title{Atkinson index}
\usage{
svyatk(formula, design, epsilon = 1, na.rm = TRUE, ...)

\method{svyatk}{survey.design}(formula, design, epsilon = 1, na.rm = TRUE, ...)

\method{svyatk}{svyrep.design}(formula, design, epsilon = 1, na.rm = TRUE, ...)

\method{svyatk}{DBIsvydesign}(formula, design, epsilon = 1, na.rm = TRUE, ...)
}
\arguments{
\item{formula}{a formula specifying the income variable.}

\item{design}{a design object of class \code{survey.design} or class \code{svyrep.design} from the \code{survey} library.}

\item{epsilon}{a parameter that determines the sensivity towards inequality in the bottom of the distribution. Defaults to epsilon = 1.}

\item{na.rm}{Should cases with missing values be dropped?}
}
\value{
Object of class "\code{cvystat}", which are vectors with a "\code{var}" attribute giving the variance and a "\code{statistic}" attribute giving the name of the statistic.
}
\description{
Estimate the Atkinson index, a measure of inequalty
}
\details{
you must run the \code{convey_prep} function on your survey design object immediately after creating it with the \code{svydesign} or \code{svrepdesign} function.
}
\examples{
library(survey)
library(vardpoor)
data(eusilc)

# linearized design
des_eusilc <- svydesign( ids = ~rb030 , strata = ~db040 ,  weights = ~rb050 , data = eusilc )

svyatk( ~eqIncome , design = des_eusilc, epsilon = .5 )

# replicate-weighted design
des_eusilc_rep <- survey:::as.svrepdesign( des_eusilc , type = "bootstrap" )
svyatk( ~eqIncome , design = des_eusilc_rep, epsilon = .5 )

# linearized design using a variable with missings
svyatk( ~eqIncome , design = des_eusilc, epsilon = .5 )
svyatk( ~eqIncome , design = des_eusilc, epsilon = .5, na.rm = TRUE )
# replicate-weighted design using a variable with missings
svyatk( ~eqIncome , design = des_eusilc_rep, epsilon = .5 )
svyatk( ~eqIncome , design = des_eusilc_rep, epsilon = .5, na.rm = TRUE )


# database-backed design
require(RSQLite)
tfile <- tempfile()
conn <- dbConnect( SQLite() , tfile )
dbWriteTable( conn , 'eusilc' , eusilc )

dbd_eusilc <- svydesign(ids = ~rb030 , strata = ~db040 ,  weights = ~rb050 , data="eusilc", dbname=tfile, dbtype="SQLite")

dbd_eusilc <- convey_prep( dbd_eusilc )
svyatk( ~eqIncome , design = dbd_eusilc, epsilon = .5 )

}
\author{
Djalma Pessoa, Anthony Damico and Guilherme Jacob
}
\references{
Matti Langel (2012). Measuring inequality in finite population sampling. PhD thesis: Universite de Neuchatel. URL \url{https://doc.rero.ch/record/29204/files/00002252.pdf}.
Martin Biewen and Stephen P. Jenkins (2002). Estimation of Generalized Entropy and Atkinson Inequality Indices from Complex Survey Data. \emph{DIW Discussion Papers}, No.345. URL \url{https://www.diw.de/documents/publikationen/73/diw_01.c.40394.de/dp345.pdf}.
}
\seealso{
\code{\link{arpr}}
}
\keyword{survey}

