% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/svygini.R
\name{svygini}
\alias{svygini}
\alias{svygini.DBIsvydesign}
\alias{svygini.survey.design}
\alias{svygini.svyrep.design}
\title{Gini coefficient}
\usage{
svygini(formula, design, ...)

\method{svygini}{survey.design}(formula, design, comp = TRUE, na.rm = FALSE,
  ...)

\method{svygini}{svyrep.design}(formula, design, na.rm = FALSE, ...)

\method{svygini}{DBIsvydesign}(x, design, ...)
}
\arguments{
\item{formula}{a formula specifying the income variable}

\item{design}{a design object of class \code{survey.design} or class \code{svyrep.design} of the library survey}

\item{comp}{logical variable \code{TRUE} if the linearized variable for domains should be completed with zeros}

\item{na.rm}{Should cases with missing values be dropped?}
}
\value{
Object of class "\code{cvystat}", which are vectors with a "\code{var}" attribute giving the variance and a "\code{statistic}" attribute giving the name of the statistic.
}
\description{
Estimate the Gini coefficient, a measure of inequalty
}
\details{
you must run the \code{convey_prep} function on your linearized survey design object immediately after creating it with the \code{svydesign} function.
}
\examples{
library(survey)
library(vardpoor)
data(eusilc)

# linearized design
des_eusilc <- svydesign( ids = ~rb030 , strata = ~db040 ,  weights = ~rb050 , data = eusilc )
des_eusilc <- convey_prep( des_eusilc )

svygini( ~eqIncome , design = des_eusilc )

# replicate-weighted design
des_eusilc_rep <- as.svrepdesign( des_eusilc , type = "bootstrap" )
des_eusilc_rep <- convey_prep( des_eusilc_rep )
svygini( ~eqIncome , design = des_eusilc_rep )

# linearized design using a variable with missings
svygini( ~ py010n , design = des_eusilc )
svygini( ~ py010n , design = des_eusilc , na.rm = TRUE )
# replicate-weighted design using a variable with missings
svygini( ~ py010n , design = des_eusilc_rep )
svygini( ~ py010n , design = des_eusilc_rep , na.rm = TRUE )
}
\author{
Djalma Pessoa and Anthony Damico
}
\references{
Guillaume Osier (2009). Variance estimation for complex indicators
of poverty and inequality. \emph{Journal of the European Survey Research
Association}, Vol.3, No.3, pp. 167-195,
ISSN 1864-3361, URL \url{http://ojs.ub.uni-konstanz.de/srm/article/view/369}.
Jean-Claude Deville (1999). Variance estimation for complex statistics and estimators:
linearization and residual techniques. Survey Methodology, 25, 193-203,
URL \url{http://www5.statcan.gc.ca/bsolc/olc-cel/olc-cel?lang=eng&catno=12-001-X19990024882}.
}
\seealso{
\code{\link{arpr}}
}
\keyword{survey}

