% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/svygpg.R
\name{svygpg}
\alias{svygpg}
\alias{svygpg.DBIsvydesign}
\alias{svygpg.survey.design}
\alias{svygpg.svyrep.design}
\title{Linearization of the gender pay (wage) gap}
\usage{
svygpg(formula, design, ...)

\method{svygpg}{survey.design}(formula, design, sex, comp = TRUE, ...)

\method{svygpg}{svyrep.design}(formula, design, sex, ...)

\method{svygpg}{DBIsvydesign}(formula, design, ...)
}
\arguments{
\item{formula}{a formula specifying the gross hourly earnings variable}

\item{design}{a design object of class \code{survey.design} or class \code{svyrep.design}
of the library survey}

\item{sex}{formula with a factor with labels 'male' and 'female'}
}
\value{
Object of class "\code{cvystat}", which are vectors with a "\code{var}" attribute giving the variance and a "\code{statistic}" attribute giving the name of the statistic.
}
\description{
Estimate the difference between the average gross hourly earnings of men and women expressed as a percentage of the average gross hourly earnings of men.
}
\examples{
set.seed(1)
y<- c(rchisq(10,10),rchisq(10,20))
H<- rep(c('str1','str2'),c(10,10))
PSU<- rep(rep(c(1,2),c(5,5)),2)
weights <- rep(2,20)

# create data frame
testset<- data.frame(y=y,H=H,psu=PSU, w=weights)
testset$sex<- rep(c(1,2),c(10,10))
testset<- transform(testset, sex= factor(sex,labels=c('female','male')))
testset$im<- 1*(testset$sex=='male')
testset$ifm<- 1*(testset$sex=='female')
testset$ym<- testset$y*(testset$sex=='male')
testset$yfm<- testset$y*(testset$sex=='female')
des<- survey::svydesign(id=~psu, strata =~H, weights =~w, data=testset, nest = TRUE )
des <- convey_prep( des )
a <-survey::svytotal(~ym+yfm+im+ifm, des)
# using the function svycontrast from the library survey
survey::svycontrast(a, quote((ym/im-yfm/im)/(ym/im)))
# using svygpg:
lin_gpg<- svygpg(~y, des, ~sex)
}
\author{
Djalma Pessoa and Anthony Damico
}
\references{
Guillaume Osier (2009). Variance estimation for complex indicators
of poverty and inequality. \emph{Journal of the European Survey Research
Association}, Vol.3, No.3, pp. 167-195,
ISSN 1864-3361, URL \url{http://ojs.ub.uni-konstanz.de/srm/article/view/369}.
Jean-Claude Deville (1999). Variance estimation for complex statistics and estimators:
linearization and residual techniques. Survey Methodology, 25, 193-203,
URL \url{http://www5.statcan.gc.ca/bsolc/olc-cel/olc-cel?lang=eng&catno=12-001-X19990024882}.
}
\seealso{
\code{\link{arpt}}
}
\keyword{survey}

