% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/svyqsr.R
\name{svyqsr}
\alias{svyqsr}
\alias{svyqsr.DBIsvydesign}
\alias{svyqsr.survey.design}
\alias{svyqsr.svyrep.design}
\title{Relative median poverty gap}
\usage{
svyqsr(formula, design, ...)

\method{svyqsr}{survey.design}(formula, design, alpha = 0.2, comp = TRUE,
  ...)

\method{svyqsr}{svyrep.design}(formula, design, alpha = 0.2, ...)

\method{svyqsr}{DBIsvydesign}(x, design, ...)
}
\arguments{
\item{formula}{a formula specifying the income variable}

\item{design}{a design object of class \code{survey.design} or class \code{svyrep.design}
of the library survey}

\item{comp}{logical variable \code{TRUE} if the inearized variable for domains
should be completed with zeros}

\item{order}{income quantile order, usually .5}

\item{percent}{fraction of the quantile, usually .60}

\item{h}{bandwidth to estimate the derivative of the cdf function}

\item{ARPR}{list generated by the function svyarpt}

\item{ncom}{length of the income vector for the whole sample}
}
\value{
a list with two components: the indicator estimate \code{value}
and the linearized variable \code{lin}.
}
\description{
Estimate the difference between the poverty threshold (arpt) and the median of  incomes less than the \code{arpt} relative to the \code{arpt}.
}
\examples{
library(vardpoor)
data(eusilc)
library(survey)
des_eusilc <- svydesign(ids=~db040, weights=~rb050, data=eusilc)
des_eusilc <- convey_prep( des_eusilc )
qsr_eqIncome<- svyqsr(~eqIncome, design=des_eusilc, alpha= .20)
}
\author{
Djalma Pessoa and Anthony Damico
}
\references{
Guillaume Osier (2009). Variance estimation for complex indicators
of poverty and inequality. \emph{Journal of the European Survey Research
Association}, Vol.3, No.3, pp. 167-195,
ISSN 1864-3361, URL \url{http://ojs.ub.uni-konstanz.de/srm/article/view/369}.
Jean-Claude Deville (1999). Variance estimation for complex statistics and estimators:
linearization and residual techniques. Survey Methodology, 25, 193-203,
URL \url{http://www5.statcan.gc.ca/bsolc/olc-cel/olc-cel?lang=eng&catno=12-001-X19990024882}.
}
\seealso{
\code{\link{arpt}}
}
\keyword{survey}

