% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svyrmir.R
\name{svyrmir}
\alias{svyrmir}
\alias{svyrmir.survey.design}
\alias{svyrmir.svyrep.design}
\title{Relative median income ratio}
\usage{
svyrmir(formula, design, ...)

\method{svyrmir}{survey.design}(formula, design, age, agelim, order = 0.5,
  na.rm = FALSE, ...)

\method{svyrmir}{svyrep.design}(formula, design, order = 0.5, age, agelim,
  na.rm = FALSE, ...)
}
\arguments{
\item{formula}{a formula specifying the income variable}

\item{design}{a design object of class \code{survey.design} or class \code{svyrep.design} of the library survey.  database-backed designs not supported}

\item{age}{formula defining the variable age}

\item{agelim}{the age cutpoint, the default is 65}

\item{order}{income quantile order, usually .5}

\item{na.rm}{Should cases with missing values be dropped?}
}
\value{
Object of class "\code{cvystat}", which are vectors with a "\code{var}" attribute giving the variance and a "\code{statistic}" attribute giving the name of the statistic.
}
\description{
Estimate the ratio between the median income of people with age above 65 and the median income of people with age below 65.
}
\details{
you must run the \code{convey_prep} function on your survey design object immediately after creating it with the \code{svydesign} or \code{svrepdesign} function.
}
\examples{
library(survey)
library(vardpoor)
data(eusilc)

# linearized design
des_eusilc <- svydesign( ids = ~rb030 , strata = ~db040 ,  weights = ~rb050 , data = eusilc )

svyrmir( ~eqIncome , design = des_eusilc , age = ~age , agelim = 65 )

# replicate-weighted design
des_eusilc_rep <- survey:::as.svrepdesign( des_eusilc , type = "bootstrap" )
svyrmir( ~eqIncome , design = des_eusilc_rep, age= ~age, agelim = 65)

# linearized design using a variable with missings
svyrmir( ~ py010n , design = des_eusilc,age= ~age, agelim = 65)
svyrmir( ~ py010n , design = des_eusilc , age= ~age, agelim = 65, na.rm = TRUE )
# replicate-weighted design using a variable with missings
svyrmir( ~ py010n , design = des_eusilc_rep,age= ~age, agelim = 65 )
svyrmir( ~ py010n , design = des_eusilc_rep ,age= ~age, agelim = 65, na.rm = TRUE )

}
\author{
Djalma Pessoa and Anthony Damico
}
\references{
Guillaume Osier (2009). Variance estimation for complex indicators
of poverty and inequality. \emph{Journal of the European Survey Research
Association}, Vol.3, No.3, pp. 167-195,
ISSN 1864-3361, URL \url{http://ojs.ub.uni-konstanz.de/srm/article/view/369}.
Jean-Claude Deville (1999). Variance estimation for complex statistics and estimators:
linearization and residual techniques. Survey Methodology, 25, 193-203,
URL \url{http://www5.statcan.gc.ca/bsolc/olc-cel/olc-cel?lang=eng&catno=12-001-X19990024882}.
}
\seealso{
\code{\link{arpt}}
}
\keyword{survey}

